prettier

+++ contracts/ReserveStewardshipIncentives.sol
@@ -11,23 +11,23 @@ contract ReserveStewardshipIncentives is IReserveStewardshipIncentives, Governab
     using ConfigHelpers for IGyroConfig;
     using FixedPoint for uint256;
 
-    uint internal constant MAX_REWARD_PERCENTAGE = 0.5e18;
-    uint internal constant OVERESTIMATION_PENALTY_FACTOR = 0.1e18;
-    uint internal constant MAX_HEALTH_VIOLATIONS = 1;  // TODO should this be configurable?
+    uint256 internal constant MAX_REWARD_PERCENTAGE = 0.5e18;
+    uint256 internal constant OVERESTIMATION_PENALTY_FACTOR = 0.1e18;
+    uint256 internal constant MAX_HEALTH_VIOLATIONS = 1; // TODO should this be configurable?
 
     /// @dev We call the collection of incentive start and end times and parameters an "initiative".
     struct Initiative {
         // SOMEDAY optimization: could be stored with fewer bits to save a slot
-        uint256 startTime;  // timestamp
-        uint256 endTime;  // timestamp
+        uint256 startTime; // timestamp
+        uint256 endTime; // timestamp
         uint256 minCollateralRatio;
         uint256 rewardPercentage;
     }
-    Initiative public activeInitiative;  // .endTime = 0 means none is there.
+    Initiative public activeInitiative; // .endTime = 0 means none is there.
 
     struct ReserveHealthViolations {
         // SOMEDAY optimization: could be stored with fewer bits to save a slot
-        uint256 lastViolatedDate;  // date
+        uint256 lastViolatedDate; // date
         uint256 nViolations;
     }
     ReserveHealthViolations public reserveHealthViolations;
@@ -42,21 +42,21 @@ contract ReserveStewardshipIncentives is IReserveStewardshipIncentives, Governab
     IGyroConfig public immutable gyroConfig;
     IGYDToken public immutable gydToken;
 
-    constructor(address _governor, address _gyroConfig) Governable(_governor)
-    {
+    constructor(address _governor, address _gyroConfig) Governable(_governor) {
         gyroConfig = IGyroConfig(_gyroConfig);
         gydToken = gyroConfig.getGYDToken();
     }
 
-    function startInitiative(uint256 rewardPercentage) external governanceOnly
-    {
+    function startInitiative(uint256 rewardPercentage) external governanceOnly {
         require(rewardPercentage <= MAX_REWARD_PERCENTAGE, "reward percentage too high");
         require(activeInitiative.endTime == 0, "active initiative already present");
 
         uint256 minCollateralRatio = gyroConfig.getStewardshipIncMinCollateralRatio();
         uint256 duration = gyroConfig.getStewardshipIncDuration();
 
-        DataTypes.ReserveState memory reserveState = gyroConfig.getReserveManager().getReserveState();
+        DataTypes.ReserveState memory reserveState = gyroConfig
+            .getReserveManager()
+            .getReserveState();
         uint256 gydSupply = gydToken.totalSupply();
 
         uint256 collateralRatio = reserveState.totalUSDValue.divDown(gydSupply);
@@ -74,7 +74,11 @@ contract ReserveStewardshipIncentives is IReserveStewardshipIncentives, Governab
             rewardPercentage: rewardPercentage
         });
         activeInitiative = initiative;
-        emit InitiativeStarted(initiative.endTime, initiative.minCollateralRatio, initiative.rewardPercentage);
+        emit InitiativeStarted(
+            initiative.endTime,
+            initiative.minCollateralRatio,
+            initiative.rewardPercentage
+        );
     }
 
     function cancelInitiative() external governanceOnly {
@@ -89,20 +93,29 @@ contract ReserveStewardshipIncentives is IReserveStewardshipIncentives, Governab
         require(initiative.endTime <= block.timestamp, "initiative not yet complete");
 
         // TODO add view methods for easier checking by others. Then use these functions, too, here.
-        
+
         // Check incentive success
-        require(reserveHealthViolations.nViolations <= MAX_HEALTH_VIOLATIONS, "initiative failed: too many health violations");
+        require(
+            reserveHealthViolations.nViolations <= MAX_HEALTH_VIOLATIONS,
+            "initiative failed: too many health violations"
+        );
 
         // Compute target reward
         uint256 gydSupply = gydToken.totalSupply();
-        uint256 aggSupply_ = aggSupply.aggSupply + (initiative.endTime - aggSupply.lastUpdatedTime) * gydSupply;
+        uint256 aggSupply_ = aggSupply.aggSupply +
+            (initiative.endTime - aggSupply.lastUpdatedTime) *
+            gydSupply;
         uint256 initiativeLength = initiative.endTime - initiative.startTime;
         uint256 avgGYDSupply = aggSupply_ / initiativeLength;
         uint256 targetReward = initiative.rewardPercentage.mulDown(avgGYDSupply);
 
         // Compute max available reward
-        DataTypes.ReserveState memory reserveState = gyroConfig.getReserveManager().getReserveState();
-        uint256 maxAllowedGYDSupply = reserveState.totalUSDValue.divDown(initiative.minCollateralRatio);
+        DataTypes.ReserveState memory reserveState = gyroConfig
+            .getReserveManager()
+            .getReserveState();
+        uint256 maxAllowedGYDSupply = reserveState.totalUSDValue.divDown(
+            initiative.minCollateralRatio
+        );
 
         require(gydSupply < maxAllowedGYDSupply, "collateral ratio too low");
         uint256 maxReward = maxAllowedGYDSupply - gydSupply;
@@ -112,7 +125,9 @@ contract ReserveStewardshipIncentives is IReserveStewardshipIncentives, Governab
         // penalty for over-estimation here.
         uint256 reward = targetReward;
         if (reward > maxReward) {
-            uint256 reduction = (FixedPoint.ONE + OVERESTIMATION_PENALTY_FACTOR).mulDown(reward - maxReward);
+            uint256 reduction = (FixedPoint.ONE + OVERESTIMATION_PENALTY_FACTOR).mulDown(
+                reward - maxReward
+            );
             reward = reduction < reward ? reward - reduction : 0;
         }
 
@@ -122,8 +137,7 @@ contract ReserveStewardshipIncentives is IReserveStewardshipIncentives, Governab
         activeInitiative.endTime = 0;
     }
 
-    function _checkpoint(DataTypes.ReserveState memory reserveState) internal
-    {
+    function _checkpoint(DataTypes.ReserveState memory reserveState) internal {
         if (activeInitiative.endTime == 0 || activeInitiative.endTime <= block.timestamp)
             // NB it's important to not track anything after the initiative has ended: may introduce manipulability.
             return;
@@ -147,8 +161,7 @@ contract ReserveStewardshipIncentives is IReserveStewardshipIncentives, Governab
         return _checkpoint(reserveState);
     }
 
-    function checkpoint() external
-    {
+    function checkpoint() external {
         DataTypes.ReserveState memory reserveState = gyroConfig
             .getReserveManager()
             .getReserveState();
@@ -157,8 +170,7 @@ contract ReserveStewardshipIncentives is IReserveStewardshipIncentives, Governab
 
     /// @dev Approximately days since epoch. Not quite correct but good enough to distinguish different days, which is
     /// all we need here.
-    function timestampToDatestamp(uint256 timestamp) internal returns (uint256)
-    {
+    function timestampToDatestamp(uint256 timestamp) internal returns (uint256) {
         return timestamp / 1 days;
     }
 }
